perf(render): isolate window chrome from workspace switch
v3.21.3 isolated AppLayout's chrome from terminal metadata/surface churn
(the workspaces array subscription), but a workspace SWITCH was a separate
path it didn't cover: every switch still re-rendered the ~1300-line chrome
(titlebar, sidebar, dock, toolbar).

Root cause was subtle. Removing AppLayout's direct activeWorkspaceId
subscription was necessary but not sufficient — useActivePaneFocus, hosted
in AppLayout, subscribes to focusKey (computeFocusKey embeds
activeWorkspaceId) to move DOM focus onto the newly active pane. That hook
re-rendering re-rendered its host, dragging the whole chrome with it.

Fix: move the switch-sensitive subscriptions out of the chrome into
render-null / viewport-scoped components.
- WorkspaceViewport now owns activeWorkspaceId/multiviewIds/paneGate/
  setActiveWorkspace/removeMultiviewWorkspace (self-subscribed, was props).
- EmptyLeafFunnel (new): the empty-leaf -> pty.create funnel effect, moved
  verbatim out of AppLayout (activeWorkspaceId + empty-leaf-key subscription).
- LayoutLogicMounts/FocusManager (new): hosts useActivePaneFocus.
- AppLayout drops those subscriptions, the funnel block, the hook call, and
  now-unused imports (-210 lines). paneGate stays (X5 discovery effect dep,
  unchanged on switch).

Measured on a live 5-workspace dev instance via direct render-count
instrumentation (component render-fn execution, not actualDuration which
includes children): chrome re-rendered 12/12 switches before, 0/12 after.
Viewport still re-renders 12/12 (the required work). Functional regression
verify: focus-follows-switch 5/5 landed on the active xterm, empty-leaf
funnel auto-created a shell for a new workspace (5->6). tsc clean.
